RCCL: net_ib_cast enabled propoer communicator hints for communicator behind GIN
