+++ projects/rccl/src/transport/net_ib_cast/connect.cc
@@ -28,32 +28,6 @@ struct ncclIbDevExtraProps {
   bool oooRq;
 };
 
-enum ncclIbCommState {
-  ncclIbCommStateStart = 0,
-  ncclIbCommStateConnect = 1,
-  ncclIbCommStateAccept = 3,
-  ncclIbCommStateSend = 4,
-  ncclIbCommStateRecv = 5,
-  ncclIbCommStateConnecting = 6,
-  ncclIbCommStateConnected = 7,
-  ncclIbCommStatePendingReady = 8,
-  ncclIbCommStateSendDevList = 9,
-  ncclIbCommStateRecvDevList = 10,
-};
-
-struct ncclIbCommStage {
-  enum ncclIbCommState state;
-  int offset;
-  void* buffer;
-  void* comm;
-};
-
-struct ncclIbHandle {
-  union ncclSocketAddress connectAddr; // Filled by the target
-  uint64_t magic; // random number to help debugging
-  int isP2p;
-  struct ncclIbCommStage stage; // Used by the other side when connecting
-};
 
 NCCL_PARAM(IbCastQpsPerConn, "IB_QPS_PER_CONNECTION", 2);
 extern int64_t rcclParamIbCastQpsPerP2p();
@@ -827,14 +801,18 @@ ib_recv_dev_list:
   comm->base.vProps = mergedDev->vProps;
   // Read isP2p from handle
   isP2p = handle->isP2p;
-  comm->useCtsOffload = IbCastIsCtsOffloadEnabled(isP2p);
+  comm->useCtsOffload = IbCastIsCtsOffloadEnabled(isP2p) && !handle->isRMA;
   if (comm->useCtsOffload) {
     comm->base.recvMatchingScheme = BY_ORDER;
   }
 
-  INFO(NCCL_NET, "NET/IB: IbCastConnect isP2p=%d", isP2p);
+  INFO(NCCL_NET, "NET/IB: IbCastConnect isP2p=%d isRMA=%d", isP2p, handle->isRMA);
   comm->base.nqps = IbCastCalculateNqps(isP2p, comm->base.vProps.ndevs, 
                                          remoteVProps.ndevs, __func__);
+  if (handle->isRMA) {
+    comm->base.nqps = 1;
+  }
+
   comm->base.nDataQps = std::max(comm->base.vProps.ndevs, remoteVProps.ndevs);
 
   if (comm->base.resiliency) {
@@ -862,6 +840,7 @@ ib_recv_dev_list:
   memset(&meta, 0, sizeof(meta));
   meta.ndevs = comm->base.vProps.ndevs;
   meta.isP2p = isP2p;
+  meta.isRMA = handle->isRMA;
 
   // Create QPs on the sender side
   NCCLCHECKGOTO(IbCastSenderQpsCreate(comm, &meta, channelId), ret, fail);
@@ -1451,8 +1430,8 @@ ib_recv:
   // Determine if Flush is enabled for this Comm. Must be done before creating
   // QPs. If Flush is enabled, extra QPs will be created for Flush operations.
   useDmaBuf  = (IbCastDmaBufSupport(lComm->dev) == ncclSuccess && ncclParamDmaBufEnable());
-  rComm->flushEnabled = ((IbCastGdrSupport() == ncclSuccess || useDmaBuf) && (!IbCastOffloadEnabled)
-                            && (ncclParamIbCastGdrFlushDisable() == 0)) ? 1 : 0;
+  rComm->flushEnabled = (((IbCastGdrSupport() == ncclSuccess || useDmaBuf) && (!IbCastOffloadEnabled)
+                            && (ncclParamIbCastGdrFlushDisable() == 0)) || remMeta.isRMA) ? 1 : 0;
 
   NCCLCHECKGOTO(IbCastReceiverQpsCreateToRts(rComm, &remMeta, &meta, channelId), ret, fail);
   if (rComm->prepostReceiveWorkRequests) {
@@ -1653,4 +1632,3 @@ ncclResult_t rcclCastNetP2pPolicy(void* handle, int isP2p) {
   ibHandle->isP2p = isP2p;
   return ncclSuccess;
 }
-
